requirements.txt
avro-python3==1.10.2
certifi==2021.10.8
charset-normalizer==2.1.1
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
